requirements.txt
streamlit>=1.28.0
openai>=1.0.0
requests>=2.28.0
pandas>=1.5.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0